pyproject.toml
[project]
name = "cmr-backend"
version = "0.1.0"
description = "CMR Backend - FastAPI starter with jobs, Modal workers"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "supabase>=2.10.0",
    "modal>=0.64.0",
    "pyjwt[crypto]>=2.9.0",
    "cryptography>=44.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
deploy_dev = "src.deployment.deploy:main"
deploy_prod = "src.deployment.deploy:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "scripts", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
